// openmetadata-ui/src/main/resources/ui/src/components/common/EntityTable/TableFilters/TableFilters.interface.ts
/*
 *  Copyright 2025 Collate.
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *  http://www.apache.org/licenses/LICENSE-2.0
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 */
import { SearchIndex } from '../../../../enums/search.enum';

export interface TableFilterOption {
  key: string;
  label: string;
  count?: number;
}

export interface TableFilterProps {
  filters: {
    owners: string[];
    glossaryTerms: string[];
    domainTypes: string[];
    tags: string[];
  };
  options: {
    owners: TableFilterOption[];
    glossaryTerms: TableFilterOption[];
    domainTypes: TableFilterOption[];
    tags: TableFilterOption[];
  };
  queryFilter?: Record<string, unknown>;
  searchIndex: SearchIndex;
  onFilterChange: (filters: TableFilterProps['filters']) => void;
  onClearAll: () => void;
}